refactor(typescript): use typescript babel profile
