+++ .babelrc
@@ -1,3 +1,3 @@
 {
-  "presets": ["@babel/preset-react"]
+  "presets": ["@babel/preset-react", "@babel/preset-typescript"]
 }
