+++ src/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './App.jsx';
+import App from './App';
 import './css/index.css';
 
 ReactDOM.render(<App />, document.getElementById('app'));
