+++ src/svg.d.ts
@@ -1,0 +1,4 @@
+declare module '*.svg' {
+  const component: React.SVGFactory;
+  export default component;
+}
