+++ tsconfig.json
@@ -1,0 +1,10 @@
+{
+    "compilerOptions": {
+        "target": "esnext",
+        "esModuleInterop": true,
+        "moduleResolution": "node",
+        "jsx": "react",
+        "noEmit": true,
+        "strict": true
+    }
+}
